pyproject.toml
[project]
name = "PyQlementine"
requires-python = ">=3.10"
version = "0.0.1"

[tool.uv]
package = false

[tool.uv.workspace]
members = ["packages/*"]

[tool.uv.sources]
PyQt6-Qlementine = { workspace = true }
PySide6-Qlementine = { workspace = true }

[dependency-groups]
pyside6 = ["PySide6-Qlementine"]
pyqt6 = ["PyQt6-Qlementine"]
test = ["pytest>=8.3.4", "pytest-qt>=4.5.0"]
dev = [
	{ "include-group" = "test" },
	"ruff>=0.9.8",
	"aqtinstall>=3.2.0",
	"prek>=0.3.1",
	"typing_extensions",
	"pyright>=1.1.408",
]

[tool.cibuildwheel]
build-verbosity = 1
build = [
	"cp310-win_amd64",
	"cp310-manylinux_x86_64",
	"cp310-macosx_x86_64",
	"cp310-macosx_arm64",
]
manylinux-x86_64-image = "manylinux_2_28"
build-frontend = "build[uv]"
test-command = "pytest {project}/tests -v"
test-requires = ["pytest>=8.3.4", "pytest-qt>=4.5.0"]

[tool.cibuildwheel.linux]
environment-pass = ["QT_VERSION"]
environment = { LLVM_INSTALL_DIR = "/usr" }
before-build = [
	"yum install -y libxkbcommon-devel libxslt clang",
	"uvx --from aqtinstall==3.2.0 aqt install-qt linux desktop $QT_VERSION --outputdir Qt --base http://mirrors.ocf.berkeley.edu/qt/",
]
repair-wheel-command = ["python scripts/repair_wheel.py {dest_dir} {wheel}"]
test-command = [
	"yum install -y epel-release",
	"yum install -y libxkbcommon-x11 xcb-util-cursor xcb-util-wm xcb-util-keysyms xorg-x11-server-Xvfb",
	"uv pip install pytest-xvfb",
	"pytest {project}/tests -v",
]

[tool.cibuildwheel.macos]
environment = { MACOSX_DEPLOYMENT_TARGET = "13.0" }
before-build = "uvx --from aqtinstall==3.2.0 aqt install-qt mac desktop $QT_VERSION --outputdir Qt --base http://mirrors.ocf.berkeley.edu/qt/"
repair-wheel-command = ["python scripts/repair_wheel.py {dest_dir} {wheel}"]

[tool.cibuildwheel.windows]
before-build = "uvx --from aqtinstall==3.2.0 aqt install-qt windows desktop %QT_VERSION% win64_msvc2022_64 --outputdir Qt --base http://mirrors.ocf.berkeley.edu/qt/"


# https://docs.astral.sh/ruff/
[tool.ruff]
line-length = 88
target-version = "py39"
src = ["src", "tests"]
fix = true
unsafe-fixes = true

[tool.ruff.lint]
pydocstyle = { convention = "numpy" }
select = [
	"E",   # style errors
	"F",   # flakes
	"W",   # warnings
	"I",   # isort
	"UP",  # pyupgrade
	"TC",  # flake8-typecheck
	"TID", # flake8-tidy-imports
	"RUF", # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["RUF012"]

# https://docs.astral.sh/ruff/formatter/
[tool.ruff.format]
docstring-code-format = true
